First batch of submodule refactorings
