+++ external/source/exploits/cve-2013-0109/nvidia_nvsvc/nvidia_nvsvc.h
@@ -1,0 +1,6 @@
+#ifndef _METASPLOIT_SOURCE_NVIDIA_NVSVC_H
+#define _METASPLOIT_SOURCE_NVIDIA_NVSVC_H
+
+VOID elevate_nvidia_nvsvc(LPVOID payload);
+
+#endif
